del1

+++ app/Http/Controllers/Module/ModuleController.php
@@ -80,7 +80,16 @@ class ModuleController extends Controller
 
     public function bulkDelete(Request $request){
         foreach ($request->ids as $id){
+            $model = $this->repository->find($id);
+            $arr = [];
+            if ($model->have_children){
+                $arr[]=$id;
+                continue;
+            }
             $this->repository->delete($id);
+        }
+        if (count ($arr) > 0){
+            return  responseJson(200, __('some items has relation cant delete'));
         }
         return  responseJson(200, __('Done'));
     }
